cleanup

+++ app/routes/checklist.$taskId_.delete/route.tsx
@@ -1,9 +1,27 @@
-import { ActionFunctionArgs, redirect } from "@remix-run/node";
+import { ActionFunctionArgs, data } from "@remix-run/node";
 import invariant from "tiny-invariant";
 import prisma from "@kiosk/audit/context/db";
 
+const taskExists = async (taskId: string) => {
+  const task = await prisma.task.findFirst({
+    where: {
+      id: taskId,
+    },
+    select: {
+      id: true,
+    },
+  });
+
+  return !!task;
+};
+
 export const action = async ({ params }: ActionFunctionArgs) => {
   invariant(params.taskId, "Missing taskId param");
+
+  const exists = await taskExists(params.taskId);
+  if (!exists) {
+    throw new Response("task:not_found", { status: 404 });
+  }
 
   await prisma.task.delete({
     where: {
@@ -11,5 +29,5 @@ export const action = async ({ params }: ActionFunctionArgs) => {
     },
   });
 
-  return redirect("/checklist");
+  return data({ success: true });
 };
